web_backEnd/src/ShopWeb.Tests/Service/ProductServiceTests.cs
using System;
using System.Collections.Generic;
using System.Threading.Tasks;
using AutoMapper;
using Moq;
using ShopWeb.Core.src.Common;
using ShopWeb.Core.src.Entity;
using ShopWeb.Core.src.RepoAbstract;
using ShopWeb.Service.src.DTOs;
using ShopWeb.Service.src.Services;
using Xunit;

namespace ShopWeb.Service.Tests
{
    public class ProductServiceTests
    {
        private readonly ProductService _productService;
        private readonly Mock<IProductRepository> _productRepositoryMock;
        private readonly Mock<IMapper> _mapperMock;

        public ProductServiceTests()
        {
            _productRepositoryMock = new Mock<IProductRepository>();
            _mapperMock = new Mock<IMapper>();

            _productService = new ProductService(_productRepositoryMock.Object, _mapperMock.Object);
        }

        [Fact]
        public async Task CreateProductAsync_ValidProductDto_ReturnsCreatedProduct()
        {
            // Arrange
            var productCreateDto = new ProductCreateDto
            {
                // Set up product create DTO
            };
            var product = new Product
            {
                // Set up product
            };

            _mapperMock.Setup(x => x.Map<Product>(productCreateDto)).Returns(product);
            _productRepositoryMock.Setup(x => x.CreateProductAsync(product)).ReturnsAsync(product);
            _mapperMock.Setup(x => x.Map<ProductReadDto>(product)).Returns(new ProductReadDto());

            // Act
            var result = await _productService.CreateProductAsync(productCreateDto);

            // Assert
            Assert.NotNull(result);
            // Add additional assertions as needed
        }

        [Fact]
        public async Task GetAllProductsAsync_ReturnsAllProducts()
        {
            // Arrange
            var queryOptions = new ProductQueryOptions
            {
                // Set up query options if necessary
            };
            var products = new List<Product>
            {
                // Set up list of products
            };
            _productRepositoryMock.Setup(x => x.GetAllProductsAsync(queryOptions)).ReturnsAsync(products);
            _mapperMock.Setup(x => x.Map<IEnumerable<ProductReadDto>>(products)).Returns(new List<ProductReadDto>());

            // Act
            var result = await _productService.GetAllProductsAsync(queryOptions);

            // Assert
            Assert.NotNull(result);
            // Add additional assertions as needed
        }

        [Fact]
        public async Task GetProductByIdAsync_ValidProductId_ReturnsProduct()
        {
            // Arrange
            var productId = Guid.NewGuid();
            var product = new Product
            {
                // Set up product
            };
            _productRepositoryMock.Setup(x => x.GetProductByIdAsync(productId)).ReturnsAsync(product);
            _mapperMock.Setup(x => x.Map<ProductReadDto>(product)).Returns(new ProductReadDto());

            // Act
            var result = await _productService.GetProductByIdAsync(productId);

            // Assert
            Assert.NotNull(result);
            // Add additional assertions as needed
        }

        [Fact]
        public async Task UpdateProductByIdAsync_ValidProductIdAndProductDto_ReturnsTrue()
        {
            // Arrange
            var productId = Guid.NewGuid();
            var productUpdateDto = new ProductUpdateDto
            {
                // Set up product update DTO
            };

            var updatedProduct = new Product
            {
                // Set up updated product
            };

            _mapperMock.Setup(x => x.Map<Product>(productUpdateDto)).Returns(updatedProduct);
            _productRepositoryMock.Setup(x => x.UpdateProductByIdAsync(productId, updatedProduct)).ReturnsAsync(true);

            // Act
            var result = await _productService.UpdateProductByIdAsync(productId, productUpdateDto);

            // Assert
            Assert.True(result);
            // Add additional assertions as needed
        }

        [Fact]
        public async Task DeleteProductByIdAsync_ValidProductId_ReturnsTrue()
        {
            // Arrange
            var productId = Guid.NewGuid();

            _productRepositoryMock.Setup(x => x.DeleteProductByIdAsync(productId)).ReturnsAsync(true);

            // Act
            var result = await _productService.DeleteProductByIdAsync(productId);

            // Assert
            Assert.True(result);
        }

        [Fact]
        public async Task GetProdyctByCategoryAsync_ValidCategoryId_ReturnsProducts()
        {
            // Arrange
            var categoryId = Guid.NewGuid();
            var products = new List<Product>
            {
                // Set up list of products
            };
            _productRepositoryMock.Setup(x => x.GetProductsByCategoryIdAsync(categoryId)).ReturnsAsync(products);
            _mapperMock.Setup(x => x.Map<IEnumerable<ProductReadDto>>(products)).Returns(new List<ProductReadDto>());

            // Act
            var result = await _productService.GetProdyctByCategoryAsync(categoryId);

            // Assert
            Assert.NotNull(result);
            // Add additional assertions as needed
        }
    }
}
